refactor: extract closeclaw-im-adapter crate (#1252)
Source: issue #1252
Type: refactor
