+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/common", "crates/config", "crates/session", "crates/llm", "crates/permission", "crates/gateway", "crates/skills", "crates/tools"]
+members = ["crates/common", "crates/config", "crates/session", "crates/llm", "crates/permission", "crates/gateway", "crates/skills", "crates/tools", "crates/im_adapter"]
 
 [workspace.package]
 edition = "2021"
@@ -94,6 +94,7 @@ closeclaw-permission = { path = "crates/permission" }
 closeclaw-gateway = { path = "crates/gateway" }
 closeclaw-skills = { path = "crates/skills" }
 closeclaw-tools = { path = "crates/tools" }
+closeclaw-im-adapter = { path = "crates/im_adapter" }
 
 # CLI argument parsing
 clap = { version = "4.4", features = ["derive"] }
