+++ crates/im_adapter/src/code_block.rs
