+++ crates/im_adapter/src/error.rs
