+++ crates/im_adapter/src/platforms/feishu/cleaner_tests.rs
