+++ src/daemon/registries.rs
@@ -137,5 +137,7 @@ async fn spawn_builtin_tools(ctx: &RegistryContext<'_>, disk_reg: &Arc<DiskSkill
         session_manager: Arc::clone(ctx.session_manager),
     });
     register_builtin_tools(ctx.tool_registry, builtin_ctx).await;
-    crate::im_adapter::platforms::feishu::tools::register_tools(ctx.tool_registry).await;
+    for tool in crate::im_adapter::platforms::feishu::tools::create_feishu_tools() {
+        ctx.tool_registry.register(tool).await.ok();
+    }
 }
